+++ packages/orq-rc/RELEASES.md
@@ -7139,3 +7139,13 @@ Based on:
 - [python v4.8.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.36 - packages/orq-rc
+
+## 2026-04-21 10:18:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.37 - packages/orq-rc